1/24기능추가 및 수정 1. 로비의 아이템 샾에서 아이템을 구매하면 게임안에서 사용할 수 있도록 개수가 동기화됨 2. 게임 안에서 아이템을 누르면 선택되고 한번 더 누르면 사용하게 됨 3. 사용하면 사용한 아이템의 개수가 하나 줄어들고 서버에도 반영됨 4. ChargeTBC로 유료재화 구매하고 UseTBC로 아이템 구매함

+++ TowerDefense/Assets/Scripts/ProjectileControl/CanonProjectile.cs
@@ -70,7 +70,7 @@ namespace ProjectileControl
             var size = Physics.OverlapSphereNonAlloc(pos, atkRange, _targetColliders, _monsterLayer);
 
             if (size <= 0) return;
-            var dividedDamage = 1.0f / size * damage;
+            var dividedDamage = (int)(1.0f / size * damage);
             for (var i = 0; i < size; i++)
             {
                 var t = _targetColliders[i];
